Updated castings for proper percentage calculations
git-svn-id: https://svn.code.sf.net/p/vpopmail/code/trunk@892 8f0cf6e2-dddb-4e5e-b901-7f5e1d474549

+++ maildirquota.c
@@ -494,7 +494,6 @@ int vmaildir_readquota(const char *dir, const char *quota_type)
    ret = client_query_quick(email, &uusage, &dusage);
    if (ret) {
 	  if (uusage != -1) {
-
 		 /*
 			Convert quota string to integers
 		 */
@@ -505,8 +504,7 @@ int vmaildir_readquota(const char *dir, const char *quota_type)
 			Return percentage
 		 */
 
-		 fd = (int)((uusage/usquota) * 100);
-
+		 fd = (int)((long double)((long double)uusage/(long double)usquota) * (long double)100.0);
 		 if (fd > 100)
 			fd = 100;
